Classic editor replacement
Add a new block that utilizes wysiwyg editor field to
replace the classic editor because a known issue with creating links.

Repurpose the button class. Making it able to register and render any
block created within Municipio.

Add a notice that is shown when required fields are not filled in.